+++ PyAPIReference/inspect_object.py
@@ -1,6 +1,7 @@
 import inspect
 import PREFS
-from types import ModuleType
+import types
+import sys
 
 def prefs(func: callable):
     """This decorator will pass the result of the given func to PREFS.convert_to_prefs, 
@@ -31,7 +32,7 @@ def prefs(func: callable):
 
     return wrapper_function # Return function to call
 
-def inspect_object(object_: object):
+def inspect_object(object_: object, exclude_types: tuple=(types.ModuleType), include_imported_members: bool=False, recursion_limit: int=10 ** 6):
 	"""Find all members of Python object.
 	Example:
 		def say_hi(name: str) -> str:
@@ -45,21 +46,24 @@ def inspect_object(object_: object):
 				default=None
 				kind=POSITIONAL_OR_KEYWORD
 			return_annotation=<class 'str'>
-	Errors:
-		Tatkes too much time, need to add some optimization.
 	"""
+	previous_recursion_limit = sys.getrecursionlimit()
+	sys.setrecursionlimit(recursion_limit)
+
 	object_name = object_.__name__
-	result = {object_name: get_object_properties(object_)}	
+	result = {object_name: get_object_properties(object_, exclude_types=exclude_types, include_imported_members=include_imported_members)}	
+	
+	sys.setrecursionlimit(previous_recursion_limit)
 
 	return result
 
-def get_object_members(object_: object, exclude_types: tuple=(ModuleType)):
+def get_object_members(object_: object, exclude_types: tuple=(types.ModuleType), include_imported_members: bool=False):
 	def filter_member(member_name: str, member: object):
 		if isinstance(member, exclude_types):
 			return False
  
  		# If the object it's a module
-		if inspect.ismodule(object_):
+		if inspect.ismodule(object_) and not include_imported_members:
 			# Get the module of the member (where it was defined or it belongs to)
 			# And check if the object name is the same as the member one.
 			# This will exclude all members that do not belong to the given module.
@@ -90,17 +94,17 @@ def get_object_members(object_: object, exclude_types: tuple=(ModuleType)):
 
 	return result
 
-def get_object_content(object_: object):
+def get_object_content(object_: object, exclude_types: tuple=(types.ModuleType), include_imported_members: bool=False):
 	"""Given an object get attributes of all of it's members.
 	"""
 	result = {}
 
-	for member_name, member in get_object_members(object_):		
-		result[member_name] = get_object_properties(member)
+	for member_name, member in get_object_members(object_, exclude_types=exclude_types, include_imported_members=include_imported_members):
+		result[member_name] = get_object_properties(member, exclude_types=exclude_types, include_imported_members=include_imported_members)
 
 	return result
 
-def get_object_properties(object_: object):
+def get_object_properties(object_: object, exclude_types: tuple=(types.ModuleType), include_imported_members: bool=False):
 	"""Given an object return it's type and content.
 	Example:
 		class Test2(Test1):
@@ -130,14 +134,14 @@ def get_object_properties(object_: object):
 	if object_type == "type":
 		object_type = "class"
 
-	result = {"type": object_type, "docstring": object_.__doc__}
+	result = {"type": object_type, "docstring": str(object_.__doc__) if object_.__doc__ is not None else object_.__doc__}
 		
 	if inspect.isclass(object_) or inspect.ismodule(object_):
 		
 		if inspect.isclass(object_):
 			result["inherits"] = [i.__name__ for i in inspect.getmro(object_)[1:-1]]
 		
-		result["content"] = get_object_content(object_)
+		result["content"] = get_object_content(object_, exclude_types=exclude_types, include_imported_members=include_imported_members)
 	
 	elif inspect.isfunction(object_) or inspect.ismethod(object_):
 		
@@ -194,9 +198,9 @@ def get_callable_parameters(callable_: callable):
 
 		
 		result[parameter.name] = {
-		"annotation": [], 
-		"default": default_parameter, 
-		"kind": parameter.kind.description
+			"annotation": [], 
+			"default": default_parameter, 
+			"kind": parameter.kind.description
 		}
 
 		if isinstance(parameter.annotation, (tuple, list)):